FE: addProducts+removeProducts - built UI: home page+control

+++ FE-React/src/slice/sliceProducts.tsx
@@ -10,17 +10,16 @@ export const getProduct = createAsyncThunk(
     async () => {
         try {
             const { data } = await instance.get(`/products`);
-            // console.log(data.data);
             return data.data;
         } catch (error) { }
     }
 );
 export const addProduct = createAsyncThunk(
     'product/addProduct',
-    async (product) => {
+    async (product:any) => {
         try {
-            const {data} = await instance.post(`/products`, product);
-            console.log(data);
+            const { data } = await instance.post(`/products`, product);
+            alert("Added products: "+ `"${product.name}"`)
             return data;
         } catch (error) { }
     }
@@ -29,7 +28,7 @@ export const updateProduct = createAsyncThunk(
     'product/updateProduct',
     async (product: any) => {
         try {
-            const {data} = await instance.patch(`/products/${product.id}`, product);
+            const { data } = await instance.patch(`/products/${product.id}`, product);
             return data;
         } catch (error) { }
     }
@@ -38,8 +37,11 @@ export const removeProduct = createAsyncThunk(
     'product/removeProducts',
     async (id) => {
         try {
-            await instance.delete(`/products/${id}`);
-            return id;
+            const conside = window.confirm("Do u want remove it !");
+            if (conside) {
+                await instance.delete(`/products/${id}`);
+                return id;
+            }
         } catch (error) { }
     }
 );
@@ -58,11 +60,11 @@ const productSlice = createSlice({
         })
         builder.addCase(updateProduct.fulfilled, (state, action) => {
             const product: any = action.payload.data
-            state.products = state.products.map((item: any) => item.id === product.id ? product : item)
+            state.products = state.products.map((item: any) => item._id == product.id ? product : item)
         })
         builder.addCase(removeProduct.fulfilled, (state, action) => {
             const id = action.payload;
-            state.products = state.products.filter((item: any) => item.id !== id)
+            state.products = state.products.filter((item: any) => item._id != id);
         })
     }
 })
